+++ CHANGELOG.md
@@ -19,3 +19,4 @@
     - Returns the chosen HTTP status code
     - Response body contains a model with both an error code and a message
   - Add duration middleware to log performance within response headers
+  - Add required Correlation ID attribute